requirements.txt
SpeechRecognition
google-genai
python-dotenv
requests
pydantic